update sendFriendRequest api, finish delete FR api, update to create relationship if the FR is accepted

+++ src/main/java/hcmute/hhkt/messengerapp/service/FriendRequestService/FriendRequestServiceImpl.java
@@ -1,5 +1,6 @@
 package hcmute.hhkt.messengerapp.service.FriendRequestService;
 
+import hcmute.hhkt.messengerapp.Exception.UnauthorizedRequestException;
 import hcmute.hhkt.messengerapp.Response.FriendRequestResponse;
 import hcmute.hhkt.messengerapp.Response.Meta;
 import hcmute.hhkt.messengerapp.Response.ResultPaginationResponse;
@@ -8,6 +9,7 @@ import hcmute.hhkt.messengerapp.domain.FriendRequest;
 import hcmute.hhkt.messengerapp.domain.User;
 import hcmute.hhkt.messengerapp.domain.enums.FriendRequestStatus;
 import hcmute.hhkt.messengerapp.repository.FriendRequestRepository;
+import hcmute.hhkt.messengerapp.service.FriendshipService.FriendshipServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,8 +19,9 @@ import java.util.UUID;
 
 @RequiredArgsConstructor
 @Service
-public class FriendRequestServiceImpl implements IFriendRequestService{
+public class FriendRequestServiceImpl implements IFriendRequestService {
     private final FriendRequestRepository friendRequestRepository;
+
     @Override
     public ResultPaginationResponse findUserOnGoingFriendRequestList(User user, Pageable pageable) {
         Page<FriendRequest> friendRequestPage = friendRequestRepository.findFriendRequestsBySender(user, pageable);
@@ -53,12 +56,23 @@ public class FriendRequestServiceImpl implements IFriendRequestService{
 
     @Override
     public FriendRequest sendFriendRequest(User senderUser, User recipientUser) {
-        if(friendRequestRepository.existsFriendRequestBySenderAndReceiver(senderUser, recipientUser) ||
-                friendRequestRepository.existsFriendRequestBySenderAndReceiver(recipientUser, senderUser)){
+        if (senderUser == recipientUser) {
+            throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_IDENTICAL_SENDER_RECEIVER);
+        }
+
+        if (friendRequestRepository.existsFriendRequestBySenderAndReceiver(senderUser, recipientUser)) {
             throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_EXIST);
         }
-        if(senderUser == recipientUser){
-            throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_IDENTICAL_SENDER_RECEIVER);
+
+        //scenario when user B rejects user A friend request
+        //later on user B sends a friend request to user A (user A cannot send another fr if it has been rejected)
+        if (friendRequestRepository.existsFriendRequestBySenderAndReceiver(recipientUser, senderUser)) {
+            FriendRequest existingFriendRequest = friendRequestRepository.findFriendRequestBySenderAndReceiver(recipientUser, senderUser);
+            if (existingFriendRequest.getStatus() == FriendRequestStatus.REJECTED) {
+                friendRequestRepository.delete(existingFriendRequest);
+            } else {
+                throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_EXIST);
+            }
         }
 
         FriendRequest newFriendRequest = FriendRequest.builder()
@@ -74,12 +88,19 @@ public class FriendRequestServiceImpl implements IFriendRequestService{
     }
 
     @Override
+    public FriendRequest findFriendRequestBySenderAndRecipient(User sender, User recipient) {
+        FriendRequest friendRequest = friendRequestRepository.findFriendRequestBySenderAndReceiver(sender, recipient);
+        return friendRequest != null ? friendRequest : friendRequestRepository.findFriendRequestBySenderAndReceiver(recipient, sender);
+    }
+
+    @Override
     public FriendRequest updateFriendRequestStatus(UUID friendRequestId, String newStatus) {
         FriendRequest updatedFriendRequest = findFriendRequestById(friendRequestId);
-        if(updatedFriendRequest == null){
+        if (updatedFriendRequest == null) {
             throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_NOT_FOUND);
         }
-        if(FriendRequestStatus.PENDING.name().equals(newStatus)){
+        if (FriendRequestStatus.PENDING.name().equals(newStatus) ||
+                updatedFriendRequest.getStatus() == FriendRequestStatus.ACCEPTED) {
             throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_INVALID_NEW_STATUS);
         }
         updatedFriendRequest.setStatus(FriendRequestStatus.valueOf(newStatus));
@@ -87,10 +108,13 @@ public class FriendRequestServiceImpl implements IFriendRequestService{
     }
 
     @Override
-    public void deleteFriendRequest(UUID friendRequestId) {
+    public void deleteFriendRequest(UUID friendRequestId, User requestedUser) {
         FriendRequest deletedFriendRequest = findFriendRequestById(friendRequestId);
-        if(deletedFriendRequest == null){
+        if (deletedFriendRequest == null) {
             throw new IllegalArgumentException(ExceptionMessage.FRIEND_REQUEST_NOT_FOUND);
+        }
+        if (requestedUser != deletedFriendRequest.getSender()) {
+            throw new UnauthorizedRequestException(ExceptionMessage.ILLEGAL_FRIEND_REQUEST_DELETE_CALLER);
         }
         friendRequestRepository.delete(deletedFriendRequest);
     }
